refactor: change to es6 lines && rename files && remove excess style constants
