--- src/assets/styles/constants/themes.scss
@@ -1,2 +1,0 @@
-$light-theme-img: url();
-$dark-theme-img: url();
