+++ src/components/CurrentDate/CurrentDate.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 
 import styles from './CurrentDate.module.scss';
-import { FORECAST_PATHS } from '../../constants/constants';
+import { FORECAST_PATHS } from '../../constants/forecaApi';
 
 function CurrentDate({ weatherImg }) {
   return (
     <div className={styles.currentDate}>
       <img
-        src={`${FORECAST_PATHS.getIcon}${weatherImg}.png`}
+        src={`${FORECAST_PATHS.getIconUrl}${weatherImg}.png`}
         alt="weather icon"
         className={styles.weatherIcon}
       />
